+++ qexpy/core/experimental_value.py
@@ -1,12 +1,27 @@
 """Defines the base class for all experimental values."""
 
+# pylint: disable=protected-access
+
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from copy import deepcopy
 from numbers import Real
+from typing import Callable
 
 import numpy as np
 
+import qexpy as q
+from qexpy.core.formula import (
+    _Formula,
+    _Add,
+    _Subtract,
+    _Multiply,
+    _Divide,
+    _Power,
+    _NegativeOp,
+    OP_TO_FORMULA,
+)
 from qexpy.utils.formatter import format_value_error
 from qexpy.utils.units import Unit
 
@@ -29,7 +44,7 @@ class ExperimentalValue(ABC):
     Examples
     --------
 
-    Values are recorded as a :func:`~qexpy.Measurement`.
+    Values are recorded as a :py:class:`~qexpy.core.Measurement`.
 
     >>> import qexpy as q
     >>> distance = q.Measurement(5, 0.1, name="distance", unit="m")
@@ -46,7 +61,7 @@ class ExperimentalValue(ABC):
     >>> print(distance.error)
     0.1
 
-    When using `ExperimentalValue` objects in calculations, the results will have properly
+    When using ``ExperimentalValue`` objects in calculations, the results will have properly
     propagated uncertainties. The units will also participate in calculations.
 
     >>> time = q.Measurement(10, 0.1, name="time", unit="s")
@@ -105,6 +120,73 @@ class ExperimentalValue(ABC):
         if isinstance(other, ExperimentalValue):
             return self.value >= other.value
         return NotImplemented
+
+    def __abs__(self):
+        if self.value < 0:
+            return -self
+        return deepcopy(self)
+
+    def __neg__(self):
+        formula = _NegativeOp(_Formula._wraps(self))
+        return q.core.DerivedValue(formula)
+
+    def __add__(self, other: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(other, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Add(_Formula._wraps(self), _Formula._wraps(other))
+        return q.core.DerivedValue(formula)
+
+    __radd__ = __add__
+
+    def __sub__(self, other: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(other, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Subtract(_Formula._wraps(self), _Formula._wraps(other))
+        return q.core.DerivedValue(formula)
+
+    def __rsub__(self, other: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(other, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Subtract(_Formula._wraps(other), _Formula._wraps(self))
+        return q.core.DerivedValue(formula)
+
+    def __mul__(self, other: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(other, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Multiply(_Formula._wraps(self), _Formula._wraps(other))
+        return q.core.DerivedValue(formula)
+
+    __rmul__ = __mul__
+
+    def __truediv__(self, other: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(other, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Divide(_Formula._wraps(self), _Formula._wraps(other))
+        return q.core.DerivedValue(formula)
+
+    def __rtruediv__(self, other: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(other, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Divide(_Formula._wraps(other), _Formula._wraps(self))
+        return q.core.DerivedValue(formula)
+
+    def __pow__(self, power: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(power, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Power(_Formula._wraps(self), _Formula._wraps(power))
+        return q.core.DerivedValue(formula)
+
+    def __rpow__(self, base: ExperimentalValue | Real) -> ExperimentalValue:
+        if not isinstance(base, (ExperimentalValue, Real)):
+            return NotImplemented
+        formula = _Power(_Formula._wraps(base), _Formula._wraps(self))
+        return q.core.DerivedValue(formula)
+
+    def __array_ufunc__(self, ufunc: Callable, _: str, *__, **___):
+        if ufunc not in OP_TO_FORMULA:
+            return NotImplemented
+        formula = OP_TO_FORMULA[ufunc](_Formula._wraps(self))
+        return q.core.DerivedValue(formula)
 
     @property
     @abstractmethod