+++ qexpy/core/measurement.py
@@ -5,6 +5,7 @@
 from __future__ import annotations
 
 import uuid
+from copy import deepcopy
 from numbers import Real
 from typing import NamedTuple, Dict, Tuple
 
@@ -13,6 +14,7 @@ import numpy as np
 import qexpy as q
 from qexpy._typing import ArrayLikeT, ArrayLike
 from .experimental_value import ExperimentalValue
+from .formula import _Formula
 
 
 class _Correlation(NamedTuple):
@@ -25,7 +27,7 @@ class _Correlation(NamedTuple):
 _correlations: Dict[Measurement, Dict[Measurement, _Correlation]] = {}
 
 
-class Measurement(ExperimentalValue):
+class Measurement(ExperimentalValue, _Formula):
     """A measurement recorded with an uncertainty
 
     Stores a measured value and the uncertainty of the measurement. A measurement can be recorded
@@ -124,6 +126,13 @@ class Measurement(ExperimentalValue):
         self._id = uuid.uuid4()
         super().__init__(name=name, unit=unit)
 
+    def __abs__(self):
+        if self.value < 0:
+            return -self
+        copy = deepcopy(self)
+        copy._id = uuid.uuid4()  # assign new id
+        return copy
+
     def __hash__(self):
         return self._id.__hash__()
 
@@ -145,6 +154,9 @@ class Measurement(ExperimentalValue):
 
         """
         return self.error
+
+    def _derivative(self, x: _Formula) -> float:
+        return 1 if self is x else 0
 
     def get_covariance(self, other: Measurement):
         """Gets the covariance between two measurements